ChatGPT_Free/prompt_3 RFW/requirements.txt
robotframework==6.1
robotframework-seleniumlibrary==6.0.0
selenium==4.10.0